QGIS tool tested
