+++ SwathProfiler/QGIS/SwathProfile.py.help
@@ -1,1 +1,1 @@
-{"Number_of_lines": "Number of lines to analyze at each side of the swath. If value is left 0, the algorithm will compute (width / cell size) lines, i.e. almost all the pixels will be taken. ", "ALG_CREATOR": "Jose Vicente Perez Pena\nDpto. Geodinamica-Universidad de Granada\n18071 Granada, Spain\nvperez@ugr.es // geolovic@gmail.com", "Draw_legend": "Flag to decide if legend will be drawn or not.", "Draw_data": "Flag to draw or not profile data as grey lines", "ALG_DESC": "This algorithm calculates the swath profiles for a set of lines for a given swath witdh. After Swath Profiles are calculated, it will show a graphic window with the output swaths. In that window use left and right mouse button to navigate between the different profiles. \n\nIt also calculates THI* index (see P\u00e9rez-Pe\u00f1a et al., 2016).\n\nP\u00e9rez-Pe\u00f1a, J.V., Al-Awabdeh, M., Aza\u00f1\u00f3n, J.M., Galve, J.P., Booth-Rea, G., Notti, D., 2016. SwathProfiler and NProfiler: Two new ArcGIS Add-ins for the automatic extraction of swath and normalized river profiles. Comput. Geosci. doi:10.1016/j.cageo.2016.08.008", "Digital_Elevation_Model": "Input Digital Elevation Model to get elevations", "Step_size": "Step size for taking elevation along each profile. If left 0, it will take elevations on each pixel. ", "Line_Shapefile": "Line shapefile with the input centerlines of the swaths. If a swath does not lie inside the DEM limits (i.e. the buffered line), that line will be discarded. Lines can be complex, but they will be simplified firts. ", "Swath_width": "Swath width. This distance will be computed AT BOTH SIDES of each polyline, i.e. the swath total width will be 2 * Swath width", "ALG_VERSION": "Version 2.0\nNovember 23, 2016\n\nLast modification December 03, 2016", "Names_field": "Field in the line shapefile that contains profile names. ", "ALG_HELP_CREATOR": "Jose Vicente Perez Pena\nDpto. Geodinamica-Universidad de Granada\n18071 Granada, Spain\nvperez@ugr.es // geolovic@gmail.com"}
+{"Number_of_lines": "Number of lines to analyze at each side of the swath. If value is left 0, the algorithm will take 50 lines at both sides of the center line. 50 lines is normally enough, but for wider swaths could be convenient select more. ", "Line_shapefile": "Input polyline shapefile with the center lines for the Swath profiles", "null": "", "ALG_CREATOR": "Jose Vicente Perez Pena\nDpto. Geodinamica-Universidad de Granada\n18071 Granada, Spain\nvperez@ugr.es // geolovic@gmail.com", "Draw_legend": "Flag to decide if legend will be drawn or not.", "Draw_data": "Flag to draw or not profile data as grey lines", "ALG_DESC": "This algorithm calculates the swath profiles for a set of lines for a given swath witdh. After Swath Profiles are calculated, it will show a graphic window with the output swaths. In that window use left and right mouse button to navigate between the different profiles. \n\nIt also calculates THI* index (see P\u00e9rez-Pe\u00f1a et al., 2016).\n\nIn the graphic windows, the following keys can be used:\n\nLEFT / RIGHT >> Select the previous / next profile\nL >> Shows / hides the legend\nT >> Shows / hides the THI graphic\nQ >> Shows / hides the q1 and q3 profiles\nE >> Select the enhanced THI (if the THI graph is visible)\nD >> Draw /hides individual profiles (can be slower)\nB >> Draw / hides a gray backgroud between max and min\n\nP\u00e9rez-Pe\u00f1a, J.V., Al-Awabdeh, M., Aza\u00f1\u00f3n, J.M., Galve, J.P., Booth-Rea, G., Notti, D., 2016. SwathProfiler and NProfiler: Two new ArcGIS Add-ins for the automatic extraction of swath and normalized river profiles. Comput. Geosci. doi:10.1016/j.cageo.2016.08.008", "Digital_Elevation_Model": "Input Digital Elevation Model to get elevations", "DEM": "Digital Elevation Model (raster)", "Width": "Half-width of the swath. The whole width of the swath will be 2 * width ", "Name_field": "Field in the line shapefile with the profile names.", "Step_size": "Step size for taking elevation along each profile. If left 0, it will take elevations on each pixel. ", "Line_Shapefile": "Line shapefile with the input centerlines of the swaths. If a swath does not lie inside the DEM limits (i.e. the buffered line), that line will be discarded. Lines can be complex, but they will be simplified firts. ", "Swath_width": "Swath width. This distance will be computed AT BOTH SIDES of each polyline, i.e. the swath total width will be 2 * Swath width", "ALG_VERSION": "Version 2.0\nNovember 23, 2016\n\nLast modification December 03, 2016", "Full_resolution": "If this option is selected, previous \u201cNumber of lines\u201d and \u201cStep size\u201d values will be ignored. The \u201cStep size\u201d will be the DEM resolution, and the \u201cNumber of lines\u201d will be width / DEM resolution. In this mode, almost all the pixels are analyzed, but can be considerably slower. ", "Names_field": "Field in the line shapefile that contains profile names. ", "ALG_HELP_CREATOR": "Jose Vicente Perez Pena\nDpto. Geodinamica-Universidad de Granada\n18071 Granada, Spain\nvperez@ugr.es // geolovic@gmail.com"}
